improved HUD and HealthBar
HealthBar on HUD fixed in position and properly updates value

+++ FPS-m/Assets/Scripts/HealthBar.cs
@@ -14,7 +14,7 @@ public class HealthBar : MonoBehaviour
         healthBar.value = playerHealth.maxHealth;
         Debug.Log("HealthBar Value: " + healthBar.value);
     }
-    public void SetHealth(int hp)
+    public void SetHealth(float hp)
     {
         healthBar.value = hp;
     }
